Add documentation and fix e2e
