+++ framework/src/streaming/examples/msk-serverless-default.lit.ts
@@ -11,6 +11,7 @@ stack.node.setContext('@data-solutions-framework-on-aws/removeDataOnDestroy', tr
 /// !show
 const msk = new MskServerless(stack, 'cluster');
 /// !hide
+
 new cdk.CfnOutput(stack, 'mskArn', {
   value: msk.cluster.attrArn,
 });
