+++ framework/src/streaming/lib/msk/msk-serverless.ts
@@ -8,6 +8,7 @@ import { CfnClusterPolicy, CfnServerlessCluster } from 'aws-cdk-lib/aws-msk';
 
 import { Construct } from 'constructs';
 import { KafkaApi } from './kafka-api';
+import { addClusterPolicy } from './msk-helpers';
 import { MskServerlessProps } from './msk-serverless-props';
 import { Authentication, MskClusterType, MskTopic, ClientAuthentication, KafkaClientLogLevel } from './msk-utils';
 import { Context, DataVpc, TrackedConstruct, TrackedConstructProps } from '../../../utils';
@@ -178,28 +179,15 @@ export class MskServerless extends TrackedConstruct {
 
 
   /**
-     * Add a cluster policy
-     *
-     * @param {PolicyDocument} policy the IAM principal to grand the consume action.
-     * @param {string} id the CDK id for the Cluster Policy
-     * @return the custom resource used to grant the consumer permissions
-     */
+    * Add a cluster policy
+    *
+    * @param {PolicyDocument} policy the IAM principal to grand the consume action.
+    * @param {string} id the CDK id for the Cluster Policy
+    * @return {CfnClusterPolicy}
+    */
   public addClusterPolicy (policy: PolicyDocument, id: string): CfnClusterPolicy {
 
-
-    let validateForResourcePolicy: string [] = policy.validateForResourcePolicy();
-
-    if (validateForResourcePolicy.length) {
-      console.log(validateForResourcePolicy.length);
-      throw new Error(`Error validating Policy document ${validateForResourcePolicy.join('\n')}`);
-    }
-
-    const cfnClusterPolicy = new CfnClusterPolicy(this, id, {
-      clusterArn: this.cluster.attrArn,
-      policy: policy.toJSON(),
-    });
-
-    return cfnClusterPolicy;
-
+    return addClusterPolicy(this, policy, id, this.cluster);
   }
+
 }
